+++ context/CDF.tsx
@@ -14,11 +14,24 @@ import React, {
 
 import { SDKConfig, CDFConfig } from "@/rainmaker.config";
 import { initCDF, CDF } from "@espressif/rainmaker-base-cdf";
-import { ESPRMBase, ESPTransportMode } from "@espressif/rainmaker-base-sdk";
+import {
+  ESPRMBase,
+  ESPRMGroup,
+  ESPRMNode,
+  ESPTransportMode,
+} from "@espressif/rainmaker-base-sdk";
+
 import { CDF_EXTERNAL_PROPERTIES } from "@/utils/constants";
 interface StoreContextType {
   store: CDF;
   isInitialized: boolean;
+  fetchAllNodes: (shouldFetchFirstPage?: boolean) => Promise<ESPRMNode[]>;
+  fetchAllGroups: (shouldFetchFirstPage?: boolean) => Promise<ESPRMGroup[]>;
+  fetchNodesAndGroups: (shouldFetchFirstPage?: boolean) => Promise<{
+    nodes: ESPRMNode[];
+    groups: ESPRMGroup[];
+  }>;
+  initUserCustomData: () => Promise<void>;
 }
 
 interface StoreProviderProps {
@@ -52,12 +65,103 @@ export const StoreProvider: React.FC<StoreProviderProps> = ({ children }) => {
     }
   };
 
+  const fetchAllNodes = async (
+    shouldFetchFirstPage: boolean = true
+  ): Promise<ESPRMNode[]> => {
+    const store = storeRef.current;
+
+    if (!store?.nodeStore) {
+      console.warn("Store or nodeStore not available");
+      return [];
+    }
+
+    try {
+      if (shouldFetchFirstPage) await store.nodeStore.syncNodeList();
+      setTimeout(async () => {
+        while (store.nodeStore.hasNext) {
+          await store.nodeStore.fetchNext();
+        }
+      }, 10);
+
+      return store.nodeStore.nodeList || [];
+    } catch (error) {
+      console.error("Error fetching nodes:", error);
+      throw error;
+    }
+  };
+
+  const fetchAllGroups = async (
+    shouldFetchFirstPage: boolean = true
+  ): Promise<ESPRMGroup[]> => {
+    const store = storeRef.current;
+
+    if (!store?.groupStore) {
+      console.warn("Store or groupStore not available");
+      return [];
+    }
+    try {
+      if (shouldFetchFirstPage) await store.groupStore.syncGroupList();
+      setTimeout(async () => {
+        while (store.groupStore.hasNext) {
+          await store.groupStore.fetchNext();
+        }
+      }, 10);
+
+      return store.groupStore.groupList || [];
+    } catch (error) {
+      console.error("Error fetching groups:", error);
+      throw error;
+    }
+  };
+
+  const fetchNodesAndGroups = async (
+    shouldFetchFirstPage: boolean = true
+  ): Promise<{
+    nodes: ESPRMNode[];
+    groups: ESPRMGroup[];
+  }> => {
+    try {
+      const [nodes, groups] = await Promise.all([
+        fetchAllNodes(shouldFetchFirstPage),
+        fetchAllGroups(shouldFetchFirstPage),
+      ]);
+
+      return { nodes, groups };
+    } catch (error) {
+      console.error("Error fetching nodes and groups:", error);
+      return { nodes: [], groups: [] };
+    }
+  };
+
+  const initUserCustomData = async () => {
+    const store = storeRef.current;
+
+    if (!store) {
+      console.warn("Custom data not initialized");
+      return;
+    }
+
+    const customData = await store.userStore.user?.getCustomData();
+    if (customData && store.userStore.userInfo) {
+      store.userStore.userInfo.customData = customData;
+    }
+  };
+
   useEffect(() => {
     initApp();
   }, []);
 
   return (
-    <StoreContext.Provider value={{ store: storeRef.current!, isInitialized }}>
+    <StoreContext.Provider
+      value={{
+        store: storeRef.current!,
+        isInitialized,
+        fetchAllNodes,
+        fetchAllGroups,
+        fetchNodesAndGroups,
+        initUserCustomData,
+      }}
+    >
       {children}
     </StoreContext.Provider>
   );